+++ public/imageupload.html
@@ -1,0 +1,17 @@
+<!-- This is a proof of concept page for uploading image files to S3 -->
+<!DOCTYPE html>
+<html lang="en">
+	<head>
+		<meta charset="UTF-8">
+		<title>Example App</title>
+	</head>
+	<body>
+
+		<form ref="uploadForm" action="/upload" method="post" enctype="multipart/form-data">
+			<input type="file" name="sampleFile"/>
+			<br><br>
+			<input type="submit" value="Upload!"/>
+		</form>
+
+	</body>
+</html>
