+++ routes/html-routes.js
@@ -13,4 +13,10 @@ module.exports = function(app) {
     res.sendFile(path.join(__dirname, "../public/main.html"));
   });
 
+
+// remove or edit when drawing functionality is added
+  app.get('/images', function(req, res) {
+    res.sendFile(path.join(__dirname, '../public/imageupload.html'));
+  });
+
 };
